+++ db/scripts/005_ddl_create_price_history_table.sql
@@ -3,5 +3,5 @@ create table price_history (
    before       bigint not null,
    after        bigint not null,
    created      timestamp without time zone default now(),
-   post_id      int references auto_post (id) not null
+   post_id      bigint references auto_post (id) not null
 );
